refactor(cli): tighten auto-update jitter review nits (test signal, init, types)
Third review round on #1600 (otReviewAgent):

- (RED) The git withdrawn-target test asserted the absence of a log substring.
  Replaced with a direct IO signal: performUpdateWithStatus's first act is
  acquireUpdateLock -> openSync(".update.lock"), so `openSyncCalls.length === 0`
  proves the updater was never entered (detect/revalidate never openSync).

- `dkg init` reruns dropped an explicit updateJitterMinutes override: a node
  pinned to `updateJitterMinutes: 0` (jitter disabled) silently reverted to the
  network/interval default on the next `dkg init`. pickPreservedAutoUpdateFields
  now preserves it (with `!== undefined` so an explicit 0 survives). Added an
  init test.

- Removed the two one-line revalidation adapters from the public daemon barrel:
  resolveCurrentNpmTarget / resolveCurrentGitTarget moved into
  auto-update-runner.ts (their only caller) as runner-scoped exports; the mapper
  tests import them from the runner module, not ../src/daemon.js.

- Runner no longer duplicates the daemon status-state shape. Exported a
  canonical `LastUpdateCheck` from daemon/state.ts and consumed it in the runner
  deps, so /api/status and the polling runner share one type.
